feat, refactor: middleware를 활용하여 serve-static를 분리하여 적용
1. src/Application에 this.server에서 정적파일을 가죠우는 로직을 middleware.run으로 수정.
2. middleware.add를 수행하는 use메소드 추가
3. middlewares/serve-static.js 생성
4. 정적파일 불러오는 함수 serveStatic구현(없으면 다음 미들웨어를
수행하도록 next()추가) 및 익스포트
5. bin.js에 serve-static임포트후 app.use()로 실행
6. bin.js에 web.html을 출력하는 함수 web 선언 후 app.use()로 실행
